ed: fix insert_line_before
the layout of struct line meant that when new->len was written, it would overwrite the ends of new->s.
for it to work with line.s declared as s[1], it would have to be at the end and then the first malloc
would allow enough space at the end of the struct for the string.

Now, the string is properly malloc'd so it has enough space wherever it is in the struct.
A bug in malloc'ing first the wrong size caused all kinds of problems!